requirements.txt
discord.py
textblob
vaderSentiment
transformers